+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/ApplicationTemplateHandler.java
@@ -20,9 +20,9 @@ import co.cask.cdap.api.templates.plugins.PluginClass;
 import co.cask.cdap.api.templates.plugins.PluginInfo;
 import co.cask.cdap.common.NotFoundException;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.internal.app.runtime.adapter.AdapterService;
 import co.cask.cdap.internal.app.runtime.adapter.ApplicationTemplateInfo;
 import co.cask.cdap.internal.app.runtime.adapter.PluginRepository;
+import co.cask.cdap.internal.app.services.AdapterService;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.template.ApplicationTemplateDetail;
 import co.cask.cdap.proto.template.ApplicationTemplateMeta;
