+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/deploy/pipeline/ApplicationTemplateVerificationStage.java
@@ -19,7 +19,7 @@ package co.cask.cdap.internal.app.deploy.pipeline;
 import co.cask.cdap.api.app.ApplicationSpecification;
 import co.cask.cdap.app.store.Store;
 import co.cask.cdap.data2.dataset2.DatasetFramework;
-import co.cask.cdap.internal.app.runtime.adapter.AdapterService;
+import co.cask.cdap.internal.app.services.AdapterService;
 import co.cask.cdap.proto.Id;
 import com.google.common.base.Preconditions;
