+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/services/ApplicationLifecycleServiceTest.java
@@ -196,70 +196,4 @@ public class ApplicationLifecycleServiceTest extends AppFabricTestBase {
       artifactRepository.clear(Id.Namespace.DEFAULT);
     }
   }
-
-  @Test
-  public void testTemplatesDeleted() throws Exception {
-    Location dummyLocation = locationFactory.create(UUID.randomUUID().toString());
-    dummyLocation.mkdirs();
-    // manually add app spec for etl batch and realtime
-    Map<String, WorkflowSpecification> workflowSpecs = ImmutableMap.of(
-      "ETLWorkflow", new WorkflowSpecification(
-        "co.cask.cdap.template.etl.batch.ETLWorkflow", "ETLWorkflow", "", Collections.<String, String>emptyMap(),
-        ImmutableList.<WorkflowNode>of(
-          new WorkflowActionNode("ETLMapReduce",
-                                 new ScheduleProgramInfo(SchedulableProgramType.MAPREDUCE, "ETLMapReduce")))
-      ));
-    Map<String, MapReduceSpecification> mrSpecs = ImmutableMap.of(
-      "ETLMapReduce", new MapReduceSpecification(
-        "co.cask.cdap.template.etl.batch.ETLMapReduce", "ETLMapReduce", "", null, null,
-        Collections.<String>emptySet(), Collections.<String, String>emptyMap(), null, null)
-    );
-    Map<String, WorkerSpecification> workerSpecs = ImmutableMap.of(
-      "ETLWorker", new WorkerSpecification(
-        "co.cask.cdap.template.etl.realtime.ETLWorker", "ETLWorker", "", Collections.<String, String>emptyMap(),
-        Collections.<String>emptySet(), null, 1)
-    );
-    ApplicationSpecification etlBatchSpec = new DefaultApplicationSpecification(
-      "ETLBatch", "", "",
-      null,
-      Collections.<String, StreamSpecification>emptyMap(),
-      Collections.<String, String>emptyMap(),
-      Collections.<String, DatasetCreationSpec>emptyMap(),
-      Collections.<String, FlowSpecification>emptyMap(),
-      mrSpecs,
-      Collections.<String, SparkSpecification>emptyMap(),
-      workflowSpecs,
-      Collections.<String, ServiceSpecification>emptyMap(),
-      Collections.<String, ScheduleSpecification>emptyMap(),
-      Collections.<String, WorkerSpecification>emptyMap(),
-      Collections.<String, Plugin>emptyMap()
-    );
-    ApplicationSpecification etlRealtimeSpec = new DefaultApplicationSpecification(
-      "ETLRealtime", "", "",
-      null,
-      Collections.<String, StreamSpecification>emptyMap(),
-      Collections.<String, String>emptyMap(),
-      Collections.<String, DatasetCreationSpec>emptyMap(),
-      Collections.<String, FlowSpecification>emptyMap(),
-      Collections.<String, MapReduceSpecification>emptyMap(),
-      Collections.<String, SparkSpecification>emptyMap(),
-      Collections.<String, WorkflowSpecification>emptyMap(),
-      Collections.<String, ServiceSpecification>emptyMap(),
-      Collections.<String, ScheduleSpecification>emptyMap(),
-      workerSpecs,
-      Collections.<String, Plugin>emptyMap()
-    );
-    Id.Application etlBatchId = Id.Application.from(Id.Namespace.DEFAULT, "ETLBatch");
-    Id.Application etlRealtimeId = Id.Application.from(Id.Namespace.DEFAULT, "ETLRealtime");
-
-    store.addApplication(etlBatchId, etlBatchSpec, dummyLocation);
-    store.addApplication(etlRealtimeId, etlRealtimeSpec, dummyLocation);
-
-    // run upgrade
-    applicationLifecycleService.upgrade(false);
-
-    // make sure apps are gone
-    Assert.assertNull(store.getApplication(etlBatchId));
-    Assert.assertNull(store.getApplication(etlRealtimeId));
-  }
 }
